Enable degug mods and FPS counter
Added some error handling for simtime when user does not define
Changed the default parameter system since the compiler flags is not
needed
Added other modes for debugging purpose. Now optimized the code for
faster operation and only update the changed cells
Do not need a temporary grid to hold the new values
Since FPS is a function of render overhead, smaller grids and smaller
debug tests may show highter FPS cap. Otherise on 8x8 it is seen as
25-28 frames per second.